merge conflict file:
   buildenv/jenkins/JenkinsfileBase
   buildenv/jenkins/openjdk_tests
   openjdk/excludes/ProblemList_openjdk8.txt

+++ get.sh
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-set -e
+set -eo pipefail
 
 SDKDIR=""
 TESTDIR="$(pwd)"
@@ -39,6 +39,7 @@ TEST_IMAGES_REQUIRED=true
 DEBUG_IMAGES_REQUIRED=true
 CURL_OPTS="s"
 CODE_COVERAGE=false
+ADDITIONAL_ARTIFACTS_REQUIRED=""
 
 usage ()
 {
@@ -146,6 +147,9 @@ parseCommandLineArgs()
 
 			"--curl_opts" )
 				CURL_OPTS="$1"; shift;;
+				
+			"--additional_artifacts_required" )
+				ADDITIONAL_ARTIFACTS_REQUIRED="$1"; shift;;
 
 			"--help" | "-h" )
 				usage; exit 0;;
@@ -192,7 +196,11 @@ getBinaryOpenjdk()
 		arr=(${download_urls/ / })
 		download_url=()
 		for n in "${arr[@]}" ; do
-			download_url+=" ${latestBuildUrl}${n}"
+			required=true
+			checkURL "$n"
+			if [ $required != false ]; then
+				download_url+=" ${latestBuildUrl}${n}"
+			fi
 		done
 	elif [ "$CUSTOMIZED_SDK_URL" != "" ]; then
 		download_url=$CUSTOMIZED_SDK_URL
@@ -254,10 +262,16 @@ getBinaryOpenjdk()
 		download_url=""
 		echo "--sdkdir is set to $SDK_RESOURCE. Therefore, skip download jdk binary"
 	fi
-
+	
 	if [ "${download_url}" != "" ]; then
 		for file in $download_url
 		do
+			if [ "$ADDITIONAL_ARTIFACTS_REQUIRED" == "RI_JDK" ]; then
+				if [[ $file = *?[0-9] ]]; then
+					fileName=$(curl -k ${curl_options} ${file}/ | grep href | sed 's/.*href="//' | sed 's/".*//' |  grep '^[a-zA-Z].*')
+					file=${file}/${fileName}
+				fi 
+			fi 
 			executeCmdWithRetry "${file##*/}" "_ENCODE_FILE_NEW=UNTAGGED curl -OLJSk${CURL_OPTS} ${curl_options} $file"
 			rt_code=$?
 			if [ $rt_code != 0 ]; then
@@ -376,7 +390,38 @@ getBinaryOpenjdk()
 					elif [[ "$jar_dir_name" =~ jre* ]] && [ "$jar_dir_name" != "j2re-image" ]; then
 						mv $jar_dir_name ../j2re-image
 					elif [[ "$jar_dir_name" =~ jdk* ]] && [ "$jar_dir_name" != "j2sdk-image" ]; then
-						mv $jar_dir_name ../j2sdk-image
+						# If test sdk has already been expanded, this one must be the additional sdk 
+						isAdditional=0
+						if [ -f "./j2sdk-image/release" ]; then 
+							isAdditional=1
+						else 
+							if [ "$ADDITIONAL_ARTIFACTS_REQUIRED" == "RI_JDK" ]; then 
+								# Check release info
+								if [ -d "./$jar_dir_name/Contents" ]; then # Mac
+									release_info=$( cat ./$jar_dir_name/Contents/Home/release )
+									UNZIPPED_ADDITIONAL_SDK="./$jar_dir_name/Contents/Home/"
+								else 	
+									release_info=$( cat ./$jar_dir_name/release )
+									UNZIPPED_ADDITIONAL_SDK="./$jar_dir_name/"
+								fi
+								if [[ "$release_info" == *"Oracle"* ]]; then
+									isAdditional=1
+								fi
+							fi
+						fi
+						if [ $isAdditional == 1 ]; then
+							if [ -d "$SDKDIR/additionaljdkbinary" ]; then
+								rm -rf $SDKDIR/additionaljdkbinary
+							else
+								mkdir $SDKDIR/additionaljdkbinary
+							fi
+							mv $UNZIPPED_ADDITIONAL_SDK/* $SDKDIR/additionaljdkbinary
+							echo "RI JDK available at $SDKDIR/additionaljdkbinary/"
+							echo "RI JDK version:"
+							$SDKDIR/additionaljdkbinary/bin/java -version
+						else 
+							mv $jar_dir_name ../j2sdk-image
+						fi
 					# The following only needed if openj9 has a different image name convention
 					elif [ "$jar_dir_name" != "j2sdk-image" ]; then
 						mv $jar_dir_name ../j2sdk-image
@@ -475,7 +520,7 @@ executeCmdWithRetry()
 			sleep $sleep_time
 
 			echo "check for $1. If found, the file will be removed."
-			if [ -f "$1" ]; then
+			if [ "$1" != "" ] && [ -f "$1" ]; then
 				echo "remove $1 before retry..."
 				rm $1
 			fi
@@ -485,8 +530,8 @@ executeCmdWithRetry()
 		rt_code=$?
 		count=$(( $count + 1 ))
 	done
-	return "$rt_code"
 	set -e
+	return "$rt_code"	
 }
 
 getFunctionalTestMaterial()
@@ -510,8 +555,7 @@ getFunctionalTestMaterial()
 	then
 		echo "update to openj9 sha: $OPENJ9_SHA"
 		cd openj9
-		echo "git fetch -q --unshallow"
-		git fetch -q --unshallow
+		executeCmdWithRetry "" "git fetch -q --unshallow"
 		if ! git checkout $OPENJ9_SHA; then
 			echo "SHA not yet found. Continue fetching PR refs and tags..."
 			echo "git fetch -q --tags $OPENJ9_REPO +refs/pull/*:refs/remotes/origin/pr/*"
@@ -636,10 +680,10 @@ testJavaVersion()
 		# Search javac as java may not be unique
 		if [[ "$CODE_COVERAGE" == "true" ]]; then
 			echo "${TEST_JDK_HOME}/build/bin/java does not exist! Searching under TEST_JDK_HOME: ${TEST_JDK_HOME}..."
-			javac_path=`find ${TEST_JDK_HOME} \( -path "*/images/jdk/bin/javac" -o -path "*/images/jdk/bin/javac.exe" \)`
+			javac_path=`find ${TEST_JDK_HOME} \( -name javac -o -name javac.exe \) | egrep '/images/jdk/bin/javac$|/images/jdk/bin/javac.exe$'`
 		else
 			echo "${TEST_JDK_HOME}/bin/java does not exist! Searching under TEST_JDK_HOME: ${TEST_JDK_HOME}..."
-			javac_path=`find ${TEST_JDK_HOME} \( -path "*/bin/javac" -o -path "*/bin/javac.exe" \)`
+			javac_path=`find ${TEST_JDK_HOME} \( -name javac -o -name javac.exe \) | egrep 'bin/javac$|bin/javac.exe$'`
 		fi
 		if [ "$javac_path" != "" ]; then
 			echo "javac_path: ${javac_path}"
